Add support for skipping encryption
