+++ src/noise/mod.rs
@@ -7,5 +7,5 @@ mod handshake;
 #[cfg(feature = "v9")]
 pub use cipher::Cipher;
 #[cfg(feature = "v10")]
-pub use cipher_v10::{segment_for_decrypt, DecryptCipher, EncryptCipher};
+pub use cipher_v10::{DecryptCipher, EncryptCipher};
 pub use handshake::{Handshake, HandshakeResult};
